♻️ refactor: Rename elo.ts to glicko.ts and center scale at DEFAULT_RATING
Align file naming with the Glicko-2 implementation and fix the
internal scale center to use DEFAULT_RATING (1200) instead of
hardcoded 1500.
